tasks, network, etc.

+++ src/host/case.c
@@ -13,15 +13,15 @@ void lowercase (char *string, int length) {
 	int i;
 
 	for (i = 0; i < length; i++)
-		string[i] = tolower (string[i]);
+		string[i] = tolower ((int) string[i]);
 }
 
 
 int w_strncmp (char *cs, char *ct, int n) {
 	int i;
 	for (i = 0; i < n; i++) {
-		if (tolower (cs[i]) != tolower (ct[i]))
-			return tolower (ct[i]) - tolower (cs[i]);
+		if (tolower ((int) cs[i]) != tolower ((int) ct[i]))
+			return tolower ((int) ct[i]) - tolower ((int) cs[i]);
 	}
 	return 0;
 }
